fix(midnight): stop using cache.freshlybakedca.ke
Midnight hosts cache.freshlybakedca.ke, so there are problems (errors)
if it tries to use it too.

Since as there's no process to filter out of a list of values without
losing the rest of the list (*except evaluating twice through scalpel
like processes), we need to disable this at the source. That means it's
best to pull this out into an ingredient and detect whether the
ingredient is enabled for turning the cache on/off...