refactor(pm/copyparty): remove useless header-hide
In the copyparty code, we previously hid a header before overriding it.

Hiding a header hides it from the client when the server sends it.
There's no point to it here. It isn't actively harmful, since as we set
the header after anyway, but it's also not required and could be
confusing.